+++ libraries/transformer/src/main/java/androidx/media3/transformer/TransformerInternal.java
@@ -21,6 +21,7 @@ import static androidx.media3.common.util.Assertions.checkState;
 import static androidx.media3.common.util.Assertions.checkStateNotNull;
 import static androidx.media3.transformer.AssetLoader.SUPPORTED_OUTPUT_TYPE_DECODED;
 import static androidx.media3.transformer.AssetLoader.SUPPORTED_OUTPUT_TYPE_ENCODED;
+import static androidx.media3.transformer.Composition.HDR_MODE_KEEP_HDR;
 import static androidx.media3.transformer.ExportException.ERROR_CODE_FAILED_RUNTIME_CHECK;
 import static androidx.media3.transformer.ExportException.ERROR_CODE_MUXING_FAILED;
 import static androidx.media3.transformer.Transformer.PROGRESS_STATE_AVAILABLE;
@@ -699,7 +700,7 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
         if (encoderFactory.videoNeedsEncoding()) {
           return true;
         }
-        if (transformationRequest.hdrMode != TransformationRequest.HDR_MODE_KEEP_HDR) {
+        if (transformationRequest.hdrMode != HDR_MODE_KEEP_HDR) {
           return true;
         }
         if (transformationRequest.videoMimeType != null