+++ libraries/transformer/src/main/java/androidx/media3/transformer/VideoSamplePipeline.java
@@ -24,10 +24,10 @@ import static androidx.media3.common.VideoFrameProcessor.INPUT_TYPE_SURFACE;
 import static androidx.media3.common.VideoFrameProcessor.INPUT_TYPE_TEXTURE_ID;
 import static androidx.media3.common.util.Assertions.checkArgument;
 import static androidx.media3.common.util.Assertions.checkNotNull;
+import static androidx.media3.transformer.Composition.HDR_MODE_KEEP_HDR;
+import static androidx.media3.transformer.Composition.HDR_MODE_TONE_MAP_HDR_TO_SDR_USING_MEDIACODEC;
+import static androidx.media3.transformer.Composition.HDR_MODE_TONE_MAP_HDR_TO_SDR_USING_OPEN_GL;
 import static androidx.media3.transformer.EncoderUtil.getSupportedEncodersForHdrEditing;
-import static androidx.media3.transformer.TransformationRequest.HDR_MODE_KEEP_HDR;
-import static androidx.media3.transformer.TransformationRequest.HDR_MODE_TONE_MAP_HDR_TO_SDR_USING_MEDIACODEC;
-import static androidx.media3.transformer.TransformationRequest.HDR_MODE_TONE_MAP_HDR_TO_SDR_USING_OPEN_GL;
 
 import android.content.Context;
 import android.graphics.Bitmap;
@@ -361,7 +361,7 @@ import org.checkerframework.dataflow.qual.Pure;
     private final TransformationRequest transformationRequest;
     private final FallbackListener fallbackListener;
     private final String requestedOutputMimeType;
-    private final @TransformationRequest.HdrMode int hdrModeAfterFallback;
+    private final @Composition.HdrMode int hdrModeAfterFallback;
 
     private @MonotonicNonNull SurfaceInfo encoderSurfaceInfo;
 
@@ -401,7 +401,7 @@ import org.checkerframework.dataflow.qual.Pure;
 
       // HdrMode fallback is only supported from HDR_MODE_KEEP_HDR to
       // HDR_MODE_TONE_MAP_HDR_TO_SDR_USING_MEDIACODEC.
-      @TransformationRequest.HdrMode int hdrMode = transformationRequest.hdrMode;
+      @Composition.HdrMode int hdrMode = transformationRequest.hdrMode;
       if (hdrMode == HDR_MODE_KEEP_HDR && isTransferHdr(inputFormat.colorInfo)) {
         ImmutableList<MediaCodecInfo> hdrEncoders =
             getSupportedEncodersForHdrEditing(requestedOutputMimeType, inputFormat.colorInfo);
@@ -422,7 +422,7 @@ import org.checkerframework.dataflow.qual.Pure;
       return Pair.create(requestedOutputMimeType, hdrMode);
     }
 
-    public @TransformationRequest.HdrMode int getHdrModeAfterFallback() {
+    public @Composition.HdrMode int getHdrModeAfterFallback() {
       return hdrModeAfterFallback;
     }
 
@@ -513,7 +513,7 @@ import org.checkerframework.dataflow.qual.Pure;
      *     processing, with {@link Format#rotationDegrees} of 90 added to the output format.
      * @param requestedFormat The requested format.
      * @param supportedFormat A format supported by the device.
-     * @param supportedHdrMode A {@link TransformationRequest.HdrMode} supported by the device.
+     * @param supportedHdrMode A {@link Composition.HdrMode} supported by the device.
      * @return The created instance.
      */
     @Pure
@@ -522,7 +522,7 @@ import org.checkerframework.dataflow.qual.Pure;
         boolean hasOutputFormatRotation,
         Format requestedFormat,
         Format supportedFormat,
-        @TransformationRequest.HdrMode int supportedHdrMode) {
+        @Composition.HdrMode int supportedHdrMode) {
       // TODO(b/259570024): Consider including bitrate in the revised fallback design.
 
       TransformationRequest.Builder supportedRequestBuilder = transformationRequest.buildUpon();